executor: split test for aggregate
