+++ executor/test/aggregate/BUILD.bazel
@@ -1,0 +1,25 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_test")
+
+go_test(
+    name = "aggregate_test",
+    timeout = "short",
+    srcs = [
+        "aggregate_test.go",
+        "main_test.go",
+    ],
+    flaky = True,
+    shard_count = 38,
+    deps = [
+        "//executor",
+        "//parser/terror",
+        "//planner/core",
+        "//session",
+        "//sessionctx/variable",
+        "//testkit",
+        "//testkit/testdata",
+        "//util/sqlexec",
+        "@com_github_pingcap_failpoint//:failpoint",
+        "@com_github_stretchr_testify//require",
+        "@org_uber_go_goleak//:goleak",
+    ],
+)
