+++ executor/test/aggregate/testdata/agg_suite_in.json
